+++ src/upstream/lib-public/generate-invoice.ts
@@ -10,6 +10,12 @@ import { generateFARR } from './FARR-generator';
 import { FaRR } from './types/FaRR.types';
 import { parseXML } from '@shared/XML-parser';
 import { i18nReady } from './i18n/i18n-init';
+import { generateCorrectivePEF } from './PEF-corrective-generator';
+import { generateBasicPEF } from './PEF-basic-generator';
+import { PEFBasicInvoice } from './types/pef-invoice.types';
+import { PEFCorrectiveInvoice } from './types/pef-invoice-corrective.types';
+import { PEFSpecInvoice } from './types/pef-invoice-spec.types';
+import { generateSpecPEF } from './PEF-spec-generator';
 
 export async function generateInvoice(
   file: File,
@@ -29,8 +35,10 @@ export async function generateInvoice(
   const xml: unknown = await parseXML(file);
 
   const wersjaFa: any = (xml as any)?.Faktura?.Naglowek?.KodFormularza?._attributes?.kodSystemowy;
+  const wersjaPef: any =
+    (xml as any)?.Invoice?.ProfileID?._text ?? (xml as any)?.CreditNote?.ProfileID?._text;
 
-  const wersja = wersjaFa;
+  const wersja = wersjaFa ?? wersjaPef;
 
   let pdf: TCreatedPdf;
 
@@ -49,6 +57,15 @@ export async function generateInvoice(
     case 'FA_RR (1)':
     case 'FA_RR(1)':
       pdf = generateFARR((xml as any).Faktura as FaRR, additionalData);
+      break;
+    case 'urn:fdc:peppol.eu:2017:poacc:billing:01:1.0':
+      pdf = generateBasicPEF((xml as any).Invoice as PEFBasicInvoice, additionalData);
+      break;
+    case 'urn:fdc:www.efaktura.gov.pl:ver2.0:corr_inv:ver4.0':
+      pdf = generateCorrectivePEF((xml as any).CreditNote as PEFCorrectiveInvoice, additionalData);
+      break;
+    case 'urn:fdc:www.efaktura.gov.pl:ver2.0:plinv:ver1.4':
+      pdf = generateSpecPEF((xml as any).Invoice as PEFSpecInvoice, additionalData);
       break;
     default:
       throw new Error(`Unknown XML Version: ${wersja}`);