+++ src/upstream/lib-public/generators/FA1/Transport.ts
@@ -53,14 +53,14 @@ export function generateTransport(transport: Transport, index?: number | null): 
         translateMap(transport.OpisLadunku, TypLadunku)
       )
     );
-    if (transport.LadunekInny?._text === '1' && transport.OpisInnegoLadunku?._text) {
-      columns.dane.push(
-        createLabelText(i18n.t('invoice.transport.cargoDescription'), i18n.t('invoice.transport.otherCargo'))
-      );
-      columns.dane.push(
-        createLabelText(i18n.t('invoice.transport.otherCargoDescription'), transport.OpisInnegoLadunku)
-      );
-    }
+  }
+  if (transport.LadunekInny?._text === '1' && transport.OpisInnegoLadunku?._text) {
+    columns.dane.push(
+      createLabelText(i18n.t('invoice.transport.cargoDescription'), i18n.t('invoice.transport.otherCargo'))
+    );
+    columns.dane.push(
+      createLabelText(i18n.t('invoice.transport.otherCargoDescription'), transport.OpisInnegoLadunku)
+    );
   }
   columns.dane.push(createLabelText(i18n.t('invoice.transport.packageUnit'), transport.JednostkaOpakowania));
   columns.dane.push(