+++ src/upstream/lib-public/i18n/i18n-init.ts
@@ -1,6 +1,6 @@
 import i18next from 'i18next';
-import pl from './lang/pl.json';
-import en from './lang/en.json';
+import { pl } from './lang/pl';
+import { en } from './lang/en';
 
 export const i18nReady: Promise<void> = i18next
   .init({