+++ src/upstream/lib-public/types/common.types.ts
@@ -1,5 +1,6 @@
 import { FP as FP3 } from './fa3.types';
 import { FP as FP2 } from './fa2.types';
+import { pl } from '../i18n/lang/pl';
 
 interface FA2FakturaZaliczkowaDataSpozaKSeF {
   NrKSeFZN: FP2;
@@ -29,3 +30,7 @@ export interface AdditionalDataTypes {
   qr2Code?: string;
   isMobile?: boolean;
 }
+
+export type TranslationKey<T> = {
+  [K in keyof T]: T[K] extends object ? TranslationKey<T[K]> : string;
+};